requirements.txt
requests~=2.32.3
jira~=3.8.0
python-dotenv~=1.0.1
protobuf~=5.29.2
pytz~=2024.2
geopy~=2.4.1
timezonefinder~=6.5.7
google-generativeai~=0.8.3
playwright~=1.49.1